pyproject.toml
[project]
name = "supermaket-service"
version = "0.1.0"
description = ""
authors = [{ name = "andre", email = "andrepedro2004@hotmail.com" }]
requires-python = ">=3.12,<4.0"
readme = "README.md"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "odmantic (>=1.0.2,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pydantic-extra-types (>=2.10.2,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "orjson (>=3.10.15,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "strawberry-graphql[debug-server,opentelemetry] (>=0.262.5,<0.263.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.52b0,<0.53)",
    "opentelemetry-instrumentation-sqlalchemy (>=0.52b0,<0.53)",
    "opentelemetry-sdk (>=1.31.0,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.31.0,<2.0.0)",
    "sqlalchemy-filterset (>=2.3.0,<3.0.0)",
    "strawberry-sqlalchemy-mapper (>=0.5.0,<0.6.0)",
    "redis[hiredis]>=4.6.0",
    "fastapi-cache2[redis]>=0.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5,<9",
    "pytest-asyncio>=0.25.3,<0.26",
    "black>=25.1.0,<26",
    "isort>=6.0.1,<7",
    "sqlalchemy-stubs>=0.4,<0.5",
    "watchfiles>=1.0.4,<2",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
